refactor: moove Warning component
